+++ app_feup/lib/controller/fetchers/print_fetcher.dart
@@ -1,6 +1,7 @@
 import 'package:uni/controller/fetchers/session_dependant_fetcher.dart';
 import 'package:uni/controller/networking/network_router.dart';
 import 'package:uni/model/entities/session.dart';
+import 'package:http/http.dart' as http;
 
 class PrintFetcher implements SessionDependantFetcher {
   @override
@@ -16,6 +17,7 @@ class PrintFetcher implements SessionDependantFetcher {
     return NetworkRouter.getWithCookies(url, query, session);
   }
 
+
   getUserPrintsMovements(Session session) {
     final String url =
         NetworkRouter.getBaseUrl('feup') + 'imp4_impressoes.logs?';
@@ -26,5 +28,27 @@ class PrintFetcher implements SessionDependantFetcher {
   getPrintHomePage(Session session) async {
     final url = 'https://print.up.pt/app?service=page/UserSummary';
     return await NetworkRouter.getWithCookies(url, {}, session);
+
+  static Future generatePrintMoneyReference(
+      double amount, Session session) async {
+    if (amount < 1.0) return Future.error('Amount less than 1,00€');
+
+    final url = NetworkRouter.getBaseUrlsFromSession(session)[0] +
+        'gpag_ccorrentes_geral.gerar_mb';
+
+    final Map data = {
+      'p_tipo_id': '3',
+      'pct_codigo': session.studentNumber,
+      'p_valor': '1',
+      'p_valor_livre': amount.toStringAsFixed(2).trim().replaceAll('.', ',')
+    };
+
+    final Map<String, String> headers = Map<String, String>();
+    headers['cookie'] = session.cookies;
+    headers['content-type'] = 'application/x-www-form-urlencoded';
+
+    final response = await http.post(url.toUri(), headers: headers, body: data);
+
+    return response;
   }
 }